requirements.txt
# Core Model & Transformers
transformers>=4.35.0
torch>=2.0.0
peft>=0.6.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Data Handling
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0

# Database & SQL
sqlparse>=0.4.4
sqlalchemy>=2.0.0
pymysql>=1.1.0

# Evaluation & Metrics
sacrebleu>=2.3.0
rouge-score>=0.1.2

# Ollama Integration
requests>=2.31.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
python-dotenv>=1.0.0
wandb>=0.15.0

# Additional dependencies
sentencepiece>=0.1.99
protobuf>=4.24.0

